ai_agent/src/requirements.txt
google-generativeai>=0.3.0
pandas>=2.2.0
openpyxl>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.29.0

